+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'apollo',
-  version: '0.3.1',
+  version: '0.4.0',
   summary: ' 🚀 Add Apollo to your Meteor app',
   git: 'https://github.com/apollostack/meteor-integration'
 });